# artifacts/api-server/src/routes/grocery.ts
import { Router, type IRouter } from "express";
import { eq } from "drizzle-orm";
import { db, groceryTable, inventoryTable } from "@workspace/db";
import {
  CreateGroceryItemBody,
  UpdateGroceryItemBody,
  UpdateGroceryItemParams,
  DeleteGroceryItemParams,
} from "@workspace/api-zod";

const router: IRouter = Router();

function computeStatus(expiryDate: string, quantity: number, lowStockThreshold: number | null): string {
  const today = new Date();
  today.setHours(0, 0, 0, 0);
  const expiry = new Date(expiryDate);
  expiry.setHours(0, 0, 0, 0);
  const diffMs = expiry.getTime() - today.getTime();
  const diffDays = Math.floor(diffMs / (1000 * 60 * 60 * 24));
  if (diffDays < 0) return "expired";
  if (lowStockThreshold != null && quantity <= lowStockThreshold) return "low_stock";
  return "fresh";
}

function toResponse(item: typeof groceryTable.$inferSelect) {
  return {
    id: item.id,
    name: item.name,
    quantity: item.quantity,
    unit: item.unit,
    category: item.category,
    isPurchased: item.isPurchased,
    isAutoGenerated: item.isAutoGenerated,
    reason: item.reason,
    addedAt: item.addedAt.toISOString(),
  };
}

router.get("/grocery", async (req, res): Promise<void> => {
  const existing = await db.select().from(groceryTable).orderBy(groceryTable.addedAt);

  const fridgeItems = await db.select().from(inventoryTable);
  const existingNames = new Set(existing.map((i) => i.name.toLowerCase()));

  const autoItems: typeof groceryTable.$inferInsert[] = [];
  for (const item of fridgeItems) {
    const status = computeStatus(item.expiryDate, item.quantity, item.lowStockThreshold);
    if (status === "low_stock" && !existingNames.has(item.name.toLowerCase())) {
      autoItems.push({
        name: item.name,
        quantity: 1,
        unit: item.unit,
        category: item.category,
        isPurchased: false,
        isAutoGenerated: true,
        reason: `Low stock — only ${item.quantity} ${item.unit} remaining`,
      });
    }
  }

  if (autoItems.length > 0) {
    await db.insert(groceryTable).values(autoItems).onConflictDoNothing();
  }

  const allItems = await db.select().from(groceryTable).orderBy(groceryTable.addedAt);
  res.json(allItems.map(toResponse));
});

router.post("/grocery", async (req, res): Promise<void> => {
  const parsed = CreateGroceryItemBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const [item] = await db.insert(groceryTable).values({
    name: parsed.data.name,
    quantity: parsed.data.quantity,
    unit: parsed.data.unit,
    category: parsed.data.category,
    reason: parsed.data.reason,
    isPurchased: false,
    isAutoGenerated: false,
  }).returning();

  res.status(201).json(toResponse(item));
});

router.put("/grocery/:id", async (req, res): Promise<void> => {
  const params = UpdateGroceryItemParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const parsed = UpdateGroceryItemBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const updateData: Partial<typeof groceryTable.$inferInsert> = {};
  if (parsed.data.name != null) updateData.name = parsed.data.name;
  if (parsed.data.quantity != null) updateData.quantity = parsed.data.quantity;
  if (parsed.data.unit != null) updateData.unit = parsed.data.unit;
  if (parsed.data.isPurchased != null) updateData.isPurchased = parsed.data.isPurchased;

  const [updated] = await db
    .update(groceryTable)
    .set(updateData)
    .where(eq(groceryTable.id, params.data.id))
    .returning();

  if (!updated) {
    res.status(404).json({ error: "Grocery item not found" });
    return;
  }

  res.json(toResponse(updated));
});

router.delete("/grocery/:id", async (req, res): Promise<void> => {
  const params = DeleteGroceryItemParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [deleted] = await db.delete(groceryTable).where(eq(groceryTable.id, params.data.id)).returning();
  if (!deleted) {
    res.status(404).json({ error: "Grocery item not found" });
    return;
  }

  res.sendStatus(204);
});

export default router;