# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Fridge & Meal Suggestion System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: inventory
    description: Fridge inventory management
  - name: recipes
    description: Recipe generation and management
  - name: grocery
    description: Grocery list management
  - name: analytics
    description: Food waste analytics
  - name: profile
    description: User profile and preferences
  - name: notifications
    description: Expiry and low-stock notifications
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /inventory:
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List all fridge inventory items
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [fresh, expiring_soon, expired, low_stock]
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add a new item to fridge inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemBody"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"

  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get a single inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          description: Item not found
    put:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemBody"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Item deleted

  /inventory/summary:
    get:
      operationId: getInventorySummary
      tags: [inventory]
      summary: Get a summary of inventory status
      responses:
        "200":
          description: Inventory summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventorySummary"

  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List all active notifications
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"

  /recipes/suggestions:
    get:
      operationId: getRecipeSuggestions
      tags: [recipes]
      summary: Get recipe suggestions based on available ingredients
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: List of suggested recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"

  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List saved recipes
      responses:
        "200":
          description: List of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
    post:
      operationId: createRecipe
      tags: [recipes]
      summary: Create or save a recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeBody"
      responses:
        "201":
          description: Recipe created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"

  /recipes/{id}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get a recipe by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Recipe not found
    delete:
      operationId: deleteRecipe
      tags: [recipes]
      summary: Delete a recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Recipe deleted

  /grocery:
    get:
      operationId: listGroceryItems
      tags: [grocery]
      summary: Get grocery list items
      responses:
        "200":
          description: List of grocery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroceryItem"
    post:
      operationId: createGroceryItem
      tags: [grocery]
      summary: Add an item to the grocery list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroceryItemBody"
      responses:
        "201":
          description: Grocery item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryItem"

  /grocery/{id}:
    put:
      operationId: updateGroceryItem
      tags: [grocery]
      summary: Update a grocery list item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroceryItemBody"
      responses:
        "200":
          description: Grocery item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryItem"
    delete:
      operationId: deleteGroceryItem
      tags: [grocery]
      summary: Remove an item from the grocery list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Grocery item removed

  /analytics:
    get:
      operationId: getAnalytics
      tags: [analytics]
      summary: Get food waste analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
            default: month
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analytics"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /substitutions:
    get:
      operationId: getSubstitutions
      tags: [recipes]
      summary: Get ingredient substitution suggestions
      parameters:
        - name: ingredient
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of substitutions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Substitution"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        category:
          type: string
          enum: [dairy, meat, vegetables, fruits, grains, condiments, beverages, frozen, other]
        expiryDate:
          type: string
          format: date
        lowStockThreshold:
          type: number
        status:
          type: string
          enum: [fresh, expiring_soon, expired, low_stock]
        addedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        daysUntilExpiry:
          type: integer
        notes:
          type: string
      required:
        - id
        - name
        - quantity
        - unit
        - category
        - expiryDate
        - status
        - addedAt
        - updatedAt
        - daysUntilExpiry

    CreateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        category:
          type: string
          enum: [dairy, meat, vegetables, fruits, grains, condiments, beverages, frozen, other]
        expiryDate:
          type: string
          format: date
        lowStockThreshold:
          type: number
        notes:
          type: string
      required:
        - name
        - quantity
        - unit
        - category
        - expiryDate

    UpdateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        category:
          type: string
          enum: [dairy, meat, vegetables, fruits, grains, condiments, beverages, frozen, other]
        expiryDate:
          type: string
          format: date
        lowStockThreshold:
          type: number
        notes:
          type: string

    InventorySummary:
      type: object
      properties:
        totalItems:
          type: integer
        freshItems:
          type: integer
        expiringSoonItems:
          type: integer
        expiredItems:
          type: integer
        lowStockItems:
          type: integer
        categoryCounts:
          type: object
          additionalProperties:
            type: integer
        urgentAlerts:
          type: array
          items:
            type: string
      required:
        - totalItems
        - freshItems
        - expiringSoonItems
        - expiredItems
        - lowStockItems
        - categoryCounts
        - urgentAlerts

    Notification:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [expiring_soon, expired, low_stock, weekly_summary]
        message:
          type: string
        itemName:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - message
        - severity
        - createdAt

    RecipeIngredient:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: string
        isOptional:
          type: boolean
        inFridge:
          type: boolean
        substitutes:
          type: array
          items:
            type: string
      required:
        - name
        - isOptional
        - inFridge

    Recipe:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        instructions:
          type: array
          items:
            type: string
        cookingTime:
          type: integer
          description: Cooking time in minutes
        servings:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
        dietaryTags:
          type: array
          items:
            type: string
        matchScore:
          type: number
          description: Percentage of required ingredients available in fridge (0-100)
        videoUrl:
          type: string
        savedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - ingredients
        - instructions
        - cookingTime
        - servings
        - difficulty
        - matchScore

    CreateRecipeBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        instructions:
          type: array
          items:
            type: string
        cookingTime:
          type: integer
        servings:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
        dietaryTags:
          type: array
          items:
            type: string
        videoUrl:
          type: string
      required:
        - name
        - ingredients
        - instructions
        - cookingTime
        - servings
        - difficulty

    GroceryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        category:
          type: string
        isPurchased:
          type: boolean
        isAutoGenerated:
          type: boolean
        reason:
          type: string
          description: Why this was added (e.g., "Low stock", "Missing for recipe")
        addedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - quantity
        - unit
        - isPurchased
        - isAutoGenerated
        - addedAt

    CreateGroceryItemBody:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        category:
          type: string
        reason:
          type: string
      required:
        - name
        - quantity
        - unit

    UpdateGroceryItemBody:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        isPurchased:
          type: boolean

    Analytics:
      type: object
      properties:
        period:
          type: string
        wastedItems:
          type: array
          items:
            $ref: "#/components/schemas/WastedItemStat"
        wastePercentage:
          type: number
          description: Percentage of food that was wasted
        estimatedMoneySaved:
          type: number
          description: Estimated money saved by using near-expiry items (in BDT)
        estimatedMoneyLost:
          type: number
          description: Estimated money lost to waste (in BDT)
        totalItemsTracked:
          type: integer
        itemsUsedBeforeExpiry:
          type: integer
        itemsExpired:
          type: integer
        topWastedCategories:
          type: array
          items:
            type: string
        monthlyTrend:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyTrendPoint"
      required:
        - period
        - wastedItems
        - wastePercentage
        - estimatedMoneySaved
        - estimatedMoneyLost
        - totalItemsTracked
        - itemsUsedBeforeExpiry
        - itemsExpired
        - topWastedCategories
        - monthlyTrend

    WastedItemStat:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        wasteCount:
          type: integer
        estimatedCost:
          type: number
      required:
        - name
        - category
        - wasteCount
        - estimatedCost

    MonthlyTrendPoint:
      type: object
      properties:
        label:
          type: string
        wastePercentage:
          type: number
        itemsTracked:
          type: integer
      required:
        - label
        - wastePercentage
        - itemsTracked

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        dietaryPreference:
          type: string
          enum: [none, halal, vegetarian, vegan]
        allergies:
          type: array
          items:
            type: string
        cookingSkillLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        availableCookingTime:
          type: integer
          description: Time available to cook in minutes
        householdSize:
          type: integer
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - dietaryPreference
        - allergies
        - cookingSkillLevel
        - availableCookingTime
        - householdSize

    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        dietaryPreference:
          type: string
          enum: [none, halal, vegetarian, vegan]
        allergies:
          type: array
          items:
            type: string
        cookingSkillLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        availableCookingTime:
          type: integer
        householdSize:
          type: integer

    Substitution:
      type: object
      properties:
        original:
          type: string
        substitute:
          type: string
        notes:
          type: string
        ratio:
          type: string
      required:
        - original
        - substitute