# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Fridge & Meal Suggestion System API
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlyTrendPoint } from "./monthlyTrendPoint";
import type { WastedItemStat } from "./wastedItemStat";

export interface Analytics {
  period: string;
  wastedItems: WastedItemStat[];
  /** Percentage of food that was wasted */
  wastePercentage: number;
  /** Estimated money saved by using near-expiry items (in BDT) */
  estimatedMoneySaved: number;
  /** Estimated money lost to waste (in BDT) */
  estimatedMoneyLost: number;
  totalItemsTracked: number;
  itemsUsedBeforeExpiry: number;
  itemsExpired: number;
  topWastedCategories: string[];
  monthlyTrend: MonthlyTrendPoint[];
}