# lib/api-zod/src/generated/types/createInventoryItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Fridge & Meal Suggestion System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateInventoryItemBodyCategory } from "./createInventoryItemBodyCategory";

export interface CreateInventoryItemBody {
  name: string;
  quantity: number;
  unit: string;
  category: CreateInventoryItemBodyCategory;
  expiryDate: Date;
  lowStockThreshold?: number;
  notes?: string;
}