# lib/api-zod/src/generated/types/wastedItemStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Fridge & Meal Suggestion System API
 * OpenAPI spec version: 0.1.0
 */

export interface WastedItemStat {
  name: string;
  category: string;
  wasteCount: number;
  estimatedCost: number;
}